Restore erased sentences on using SSH verification.
Mention SSH protocol and link to GitHub directions for generating SSH
keys.

I had contacted Authorea, but they weren't sure what could have caused
this problem. Here is the description I sent them:

I stumbled across a strange situation today. I was looking for a
change I knew I had made to a file becuase it was in the log, but it
turns out it was deleted in a subsequent commit without my knowledge.

On July 30th I made the following edit via the Authorea site:

https://github.com/jdblischak/git-for-science/commit/592a58c15ca3efd7b5b45ac0c69e860f81ddad88
https://www.authorea.com/users/5990/articles/17489/commits/592a58c15ca3efd7b5b45ac0c69e860f81ddad88

Then the next day, I made another change to the file, again using the
Authorea site. But for some reason, it deleted the change I had made
the previous day:

https://github.com/jdblischak/git-for-science/commit/a9b8382b4e6a13553d10c95b0162320734c289c7
https://www.authorea.com/users/5990/articles/17489/commits/a9b8382b4e6a13553d10c95b0162320734c289c7